% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dozen.R
\name{dozen}
\alias{dozen}
\title{Multiply the number from 1 to 12}
\usage{
dozen(b)
}
\arguments{
\item{b}{the number to multiply by 12}
}
\value{
the result should be multiples of a number from 1 to 12
}
\description{
This package simply takes the number and multiplies it from 1 to 12
The result simply shows 1 * 2 = 2
}
\examples{
2 * 1 = 2
2 * 2 = 4
2 * 3 = 6
2 * 4 = 8
2 * 5 = 10
2 * 6 = 12
2 * 7 = 14
2 * 8 = 16
2 * 9 = 18
2 * 10 = 20
2 * 11 = 22
2 * 12 = 24
}
